Creates webpack plugins module
